+++ src/scene/Light.h
@@ -71,7 +71,7 @@ extern long MAX_LLIGHTS;
 extern EERIE_LIGHT * PDL[MAX_DYNLIGHTS];
 extern EERIE_LIGHT * GLight[MAX_LIGHTS];
 extern EERIE_LIGHT DynLight[MAX_DYNLIGHTS];
-extern long TOTPDL;
+extern size_t TOTPDL;
 extern size_t TOTIOPDL;
 
 ARX_HANDLE_TYPEDEF(long, LightHandle, -1)
